huggingface-deploy/backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlmodel==0.0.14
psycopg2-binary==2.9.9
pydantic-settings==2.1.0
python-dotenv==1.0.0
pytest==7.4.4
httpx==0.26.0
pytest-cov==4.1.0

# Phase III: AI Chatbot
# HuggingFace Inference API (recommended - no local model needed)
httpx==0.26.0

# Local model support (optional - fallback if HF API not available)
transformers==4.37.0
torch==2.2.0
accelerate==0.27.0
sentencepiece==0.1.99
